Replace for..in with for..of to save 6 bytes
This replaces the for..in-loop in the unpacking bootstrapper with a
for..of loop. This shaves 6 bytes off of the total packed output.

The for..of loop is supported by all current major browsers: Chrome
since version 51, Firefox since version 13, Edge since version 12,
Safari since version 7.1, Chrome for Android since version 38 and Safari
Mobile since version 8.